+++ src/main/kotlin/interpreter/Beatwalls.kt
@@ -1,21 +1,45 @@
 package interpreter
 
 import beatwalls.logError
+import beatwalls.logInfo
 import beatwalls.mainFileSuffix
 import interpreter.parser.Parser
+import kotlinx.coroutines.cancel
+import kotlinx.coroutines.runBlocking
+import structure.bwElements.BwElement
 import java.io.File
 
 
-class Beatwalls(val workingDirectory: File) {
+class Beatwalls(val workingDirectory: File, val bwFiles: List<File>) {
     var hadError = false
     val options = Options(workingDirectory)
-    val mainFile = File(workingDirectory, mainFileSuffix)
 
     fun run(){
-        hadError = false
 
+        hadError = false
         val ml = try { MapLoader(workingDirectory) }catch (e: Exception){ return }
-        val s = Scanner(mainFile.readText(), this, mainFile)
+
+        val elements = mutableListOf<BwElement>()
+        for (currentFile in bwFiles){
+                logInfo("Adding elements from ${currentFile.name}")
+                val l = loadElements(currentFile)?: return
+                elements.addAll(l)
+        }
+
+        logInfo("Translating ${elements.size} elements to ingame Objects")
+        val tr = Translator(elements, this)
+        tr.translate()
+        val diff = ml.loadDifficulty(options.Characteristic, options.difficulty)?: return
+
+        logInfo("Correcting timing to bmp changes")
+        val bpmAdjuster = BpmAdjuster(diff)
+        bpmAdjuster.correctElements(tr.obst, tr.notes, tr.events)
+
+        ml.writeDiff(diff,tr.obst,tr.notes, tr.events)
+    }
+
+    fun loadElements(file: File): List<BwElement>? {
+        val s = Scanner(file.readText(), this, file)
         val blocks = s.scan()
 
         val p = Parser(blocks, this)
@@ -23,20 +47,12 @@ class Beatwalls(val workingDirectory: File) {
 
         if (hadError){
             logError("Looks like you have some Errors. Come back if you fixed them")
-            return
+            return null
         }
 
         val ev = Evaluator(structs, this)
-        val elements = ev.evaluate()
+        return ev.evaluate()
 
-        val tr = Translator(elements, this)
-        tr.translate()
-        val diff = ml.loadDifficulty(options.Characteristic, options.difficulty)?: return
-
-        val bpmAdjuster = BpmAdjuster(diff)
-        bpmAdjuster.correctElements(tr.obst, tr.notes, tr.events)
-
-        ml.writeDiff(diff,tr.obst,tr.notes, tr.events)
     }