+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 You can learn more [here](https://spooky.moe/beatwalls).
 
 Change Log:
+- Update Include Syntax: They must be at the top of a file
+- Update Include Scope: Custom Structures and Interfaces are now **private** per file